+++ packages/integration-sdk-runtime/src/storage/FileSystemGraphObjectStore/__tests__/FileSystemGraphObjectStore.test.ts
@@ -37,6 +37,7 @@ async function getStorageDirectoryDataForStep(
   const accumulatedWrittenStepData: FlushedGraphObjectData = {
     entities: [],
     relationships: [],
+    mappedRelationships: [],
   };
 
   for (const collectionType of ['entities', 'relationships']) {