+++ packages/integration-sdk-runtime/src/storage/types.ts
@@ -1,4 +1,8 @@
-import { Entity, Relationship } from '@jupiterone/integration-sdk-core';
+import {
+  Entity,
+  MappedRelationship,
+  Relationship,
+} from '@jupiterone/integration-sdk-core';
 
 export interface FlushedEntityData {
   entities: Entity[];
@@ -8,5 +12,10 @@ export interface FlushedRelationshipData {
   relationships: Relationship[];
 }
 
+export interface FlushedMappedRelationshipData {
+  mappedRelationships: MappedRelationship[];
+}
+
 export type FlushedGraphObjectData = FlushedEntityData &
-  FlushedRelationshipData;
+  FlushedRelationshipData &
+  FlushedMappedRelationshipData;