+++ packages/integration-sdk-runtime/src/synchronization/__tests__/uploadData.test.ts
@@ -1,7 +1,14 @@
 import times from 'lodash/times';
 import noop from 'lodash/noop';
 
-import { Entity, Relationship } from '@jupiterone/integration-sdk-core';
+import {
+  createIntegrationEntity,
+  createMappedRelationship,
+  Entity,
+  MappedRelationship,
+  Relationship,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
 import { generateSynchronizationJob } from './util/generateSynchronizationJob';
 
 import { getApiBaseUrl, createApiClient } from '../../api';
@@ -99,6 +106,67 @@ test('uploads relationship data in batches of 250', async () => {
     `/persister/synchronization/jobs/${job.id}/relationships`,
     {
       relationships: data.slice(500, 510),
+    },
+  );
+});
+
+test('uploads mapped relationship data in batches of 250', async () => {
+  const { job, logger, apiClient } = createTestContext();
+
+  const data = times(
+    510,
+    (i): MappedRelationship =>
+      createMappedRelationship({
+        _class: RelationshipClass.HAS,
+        _type: 'relationship_resource',
+        _key: `mapped-relationship:${i}`,
+        source: createIntegrationEntity({
+          entityData: {
+            source: {},
+            assign: {
+              _class: 'Resource',
+              _type: 'entity_resource',
+              _key: `entity:${i}`,
+            },
+          },
+        }),
+        target: {
+          _type: 'entity_resource',
+          _key: `entity:${i + 1}`,
+        },
+      }),
+  );
+
+  const postSpy = jest.spyOn(apiClient, 'post').mockImplementation(noop as any);
+
+  await uploadData(
+    {
+      job,
+      logger,
+      apiClient,
+    },
+    'mapped-relationships',
+    data,
+  );
+
+  expect(postSpy).toHaveBeenCalledTimes(3);
+
+  expect(postSpy).toHaveBeenCalledWith(
+    `/persister/synchronization/jobs/${job.id}/mapped-relationships`,
+    {
+      'mapped-relationships': data.slice(0, 250),
+    },
+  );
+  expect(postSpy).toHaveBeenCalledWith(
+    `/persister/synchronization/jobs/${job.id}/mapped-relationships`,
+    {
+      'mapped-relationships': data.slice(250, 500),
+    },
+  );
+  expect(postSpy).toHaveBeenCalledWith(
+    `/persister/synchronization/jobs/${job.id}/mapped-relationships`,
+    {
+      'mapped-relationships': data.slice(500, 510),
     },
   );
 });
@@ -238,6 +306,7 @@ test('should not upload with empty graph object arrays', async () => {
     {
       entities: [],
       relationships: [],
+      mappedRelationships: [],
     },
   );